toolchain path flags and refactoring
